Setup
just moves stuff around a bit
